% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02buka_Func.R
\name{buKa_NameTransfer}
\alias{buKa_NameTransfer}
\title{针对补卡数据进行处理}
\usage{
buKa_NameTransfer(data)
}
\arguments{
\item{data}{deal_buKa得到的数据}
}
\value{
返回值
}
\description{
针对补卡数据进行处理
}
\examples{
buKa_NameTransfer();
}
