% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01checkOn_Func.R
\name{checkOn_all}
\alias{checkOn_all}
\title{考勤数据入口函数}
\usage{
checkOn_all(file = "data-raw/input/01-ns_kaoQin.xlsx",
  morningStartTime = "0:00:00", moringEndTime = "6:00:00",
  AmStartTime = "8:00:00", AmCheckTime = "9:00:00",
  AmEndTime = "12:00:00", AmLateMins = 0, pmStartTime = "12:00:00",
  pmCheckTime = "18:00:00", pmEndTime = "18:30:00", PmLeakMins = 0,
  nightStartTime = "18:30:00", nightCheckTime = "19:00:00")
}
\arguments{
\item{file}{文件名}

\item{morningStartTime}{凌晨开始时间}

\item{moringEndTime}{凌晨结束时间}

\item{AmStartTime}{上午开始}

\item{AmCheckTime}{上午打卡}

\item{AmEndTime}{上午结束}

\item{AmLateMins}{上午偏置}

\item{pmStartTime}{下午开始}

\item{pmCheckTime}{下午打卡}

\item{pmEndTime}{下午结束}

\item{PmLeakMins}{下午偏置}

\item{nightStartTime}{晚上开始}

\item{nightCheckTime}{晚上打卡}
}
\value{
返回值
}
\description{
考勤数据入口函数
}
\examples{
CheckOn_all();
}
