% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01checkOn_Func.R
\name{checkOn_AM}
\alias{checkOn_AM}
\title{提供数据中的早上数据}
\usage{
checkOn_AM(data_CheckOn, AmCheckTime = "9:00:00", AmLateMins = 5)
}
\arguments{
\item{data_CheckOn}{考勤数据}

\item{AmCheckTime}{上班打卡时间}

\item{AmLateMins}{迟到分钟数}
}
\value{
返回值
}
\description{
提供数据中的早上数据
}
\examples{
checkOn_AM()
}
