% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01checkOn_Func.R
\name{checkOn_sum}
\alias{checkOn_sum}
\title{针对数据进行汇总处理}
\usage{
checkOn_sum(checkOn_all_data, startDate = "2019-06-01",
  EndDate = "2019-06-30", ErrorOnly = FALSE)
}
\arguments{
\item{checkOn_all_data}{考勤数据}

\item{startDate}{月度开始}

\item{EndDate}{月度结束}

\item{ErrorOnly}{是否只统计异常数据}
}
\value{
返回值
}
\description{
针对数据进行汇总处理
}
\examples{
checkOn_sum();
}
