% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01checkOn_Func.R
\name{checkOn_sum_byDay}
\alias{checkOn_sum_byDay}
\title{按天计算汇总数据}
\usage{
checkOn_sum_byDay(checkOn_all_data, startDate = "2019-06-01",
  EndDate = "2019-06-30", ErrorOnly = FALSE)
}
\arguments{
\item{checkOn_all_data}{考勤数据}

\item{startDate}{开始日期}

\item{EndDate}{结束日期}

\item{ErrorOnly}{是否只包含错误}
}
\value{
返回值
}
\description{
按天计算汇总数据
}
\examples{
checkOn_sum_byDay();
}
