% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03Achuchai_func.R
\name{chuChai_NameTransfer}
\alias{chuChai_NameTransfer}
\title{针对出差数据进行名称转换}
\usage{
chuChai_NameTransfer(data)
}
\arguments{
\item{data}{出差deal_chuChai处理后数据}
}
\value{
返回值
}
\description{
针对出差数据进行名称转换
}
\examples{
chuChai_NameTransfer();
}
