% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03Achuchai_func.R
\name{chuChai_splitIntoMultiRows}
\alias{chuChai_splitIntoMultiRows}
\title{针对出差数据进行处理}
\usage{
chuChai_splitIntoMultiRows(data, startDateName = "FStartDate",
  endDateName = "FEndDate")
}
\arguments{
\item{data}{数据}

\item{startDateName}{日期}

\item{endDateName}{结束日期}
}
\value{
返回值
}
\description{
针对出差数据进行处理
}
\examples{
chuChai_splitIntoMultiRows();
}
