% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03Achuchai_func.R
\name{chuChai_sum}
\alias{chuChai_sum}
\title{针对出差数据进行处理}
\usage{
chuChai_sum(data, startDate = "2019-07-01", endDate = "2019-07-31")
}
\arguments{
\item{data}{出差明细数据}

\item{startDate}{开始日期}

\item{endDate}{结束日期}
}
\value{
返回值
}
\description{
针对出差数据进行处理
}
\examples{
chuChai_sum();
}
