% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03Achuchai_func.R
\name{chuChai_sum_byDay}
\alias{chuChai_sum_byDay}
\title{出差数据按天汇总}
\usage{
chuChai_sum_byDay(data, startDate = "2019-07-01",
  endDate = "2019-07-31")
}
\arguments{
\item{data}{数据}

\item{startDate}{开始日期}

\item{endDate}{结束日期}
}
\value{
返回值
}
\description{
出差数据按天汇总
}
\examples{
chuChai_sum_byDay();
}
