% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02buka_Func.R
\name{deal_buKa}
\alias{deal_buKa}
\title{网商补卡记录}
\usage{
deal_buKa(file = "data-raw/input/02-ns_buKa.xlsx",
  AmCheckTime = "9:00:00", pmCheckTime = "18:00:00")
}
\arguments{
\item{file}{补卡钉钉导出文件}
}
\value{
返回值
}
\description{
网商补卡记录
}
\examples{
deal_buKa();
}
