% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01checkOn_Func.R
\name{deal_CheckOn}
\alias{deal_CheckOn}
\title{处理网商考勤数据}
\usage{
deal_CheckOn(file = "data-raw/input/01-ns_kaoQin.xlsx",
  morningStartTime = "0:00:00", moringEndTime = "6:00:00",
  AmStartTime = "8:00:00", AmCheckTime = "9:00:00",
  AmEndTime = "12:00:00", pmStartTime = "12:00:00",
  pmCheckTime = "18:00:00", pmEndTime = "18:30:00",
  nightStartTime = "18:30:00", nightCheckTime = "19:00:00")
}
\arguments{
\item{file}{考勤数据文件}

\item{morningStartTime}{凌晨开始时点}

\item{moringEndTime}{凌晨结束时点}

\item{AmStartTime}{上午开始时间}

\item{AmCheckTime}{上午打卡时间}

\item{AmEndTime}{上午结束时间}

\item{pmStartTime}{下午开始时间}

\item{pmCheckTime}{下午打卡时间}

\item{pmEndTime}{下午结束时间}

\item{nightStartTime}{晚上开始时间}

\item{nightCheckTime}{晚上结束时间}
}
\value{
返回数据框
}
\description{
处理网商考勤数据
}
\examples{
deal_CheckOn();
}
