% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03Achuchai_func.R
\name{deal_chuChai}
\alias{deal_chuChai}
\title{处理网商出差数据}
\usage{
deal_chuChai(file = "data-raw/input/03-ns_chuchai.xlsx")
}
\arguments{
\item{file}{文件名}
}
\value{
返回值
}
\description{
处理网商出差数据
}
\examples{
deal_chuChai();
}
