% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/04jiaban_Func.R
\name{deal_jiaBan}
\alias{deal_jiaBan}
\title{处理网商的加班记录}
\usage{
deal_jiaBan(file = "data-raw/input/04-ns_jiaban.xlsx")
}
\arguments{
\item{file}{文件名}
}
\value{
返回值
}
\description{
处理网商的加班记录
}
\examples{
deal_jiaBan();
}
