% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/05qingjia_Func.R
\name{deal_qingJia}
\alias{deal_qingJia}
\title{处理请假业务}
\usage{
deal_qingJia(file = "data-raw/input/05-ns_qingjia.xlsx")
}
\arguments{
\item{file}{文件名}
}
\value{
返回值
}
\description{
处理请假业务
}
\examples{
deal_qingJia();
}
