% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03Bwaichu_Func.R
\name{deal_waiChu}
\alias{deal_waiChu}
\title{处理外出数据}
\usage{
deal_waiChu(file = "data-raw/input/06-ns_waichu.xlsx")
}
\arguments{
\item{file}{文件名}
}
\value{
返回值
}
\description{
处理外出数据
}
\examples{
deal_waiChu();
}
