% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/04jiaban_Func.R
\name{jiaBan_NameTransfer}
\alias{jiaBan_NameTransfer}
\title{处理用于加班时间的统计}
\usage{
jiaBan_NameTransfer(data)
}
\arguments{
\item{data}{deal_jiaBan处理后的数据}
}
\value{
返回值
}
\description{
处理用于加班时间的统计
}
\examples{
jiaBan_NameTransfer();
}
