% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/04jiaban_Func.R
\name{jiaBan_sum}
\alias{jiaBan_sum}
\title{处理加班汇总数据}
\usage{
jiaBan_sum(data, startDate = "2019-07-01", endDate = "2019-07-31")
}
\arguments{
\item{data}{加班明细数据}

\item{startDate}{开始日期}

\item{endDate}{结束日期}
}
\value{
返回值
}
\description{
处理加班汇总数据
}
\examples{
jiaBan_sum();
}
