% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{ns_Calendar}
\alias{ns_Calendar}
\title{获取网商日历}
\usage{
ns_Calendar(file = "data-raw/public/ns_calendar.xlsx")
}
\arguments{
\item{file}{文件名}
}
\value{
返回值
}
\description{
获取网商日历
}
\examples{
ns_Calendar();
}
