% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{ns_NameTable}
\alias{ns_NameTable}
\title{提供名称对照表}
\usage{
ns_NameTable(file = "data-raw/public/ns_NameTable.xlsx",
  sheet = "mapping")
}
\arguments{
\item{file}{对照表文件名}

\item{sheet}{页签mapping}
}
\value{
返回值
}
\description{
提供名称对照表
}
\examples{
ns_NameTable();
}
