% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{nshr}
\alias{nshr}
\title{用于处理nshr文件}
\usage{
nshr(file, base_dir = "./data-raw/input/")
}
\arguments{
\item{file}{文件名称}

\item{base_dir}{基本路径}
}
\value{
返回值
}
\description{
用于处理nshr文件
}
\examples{
nshr();
}
