% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/05qingjia_Func.R
\name{qingJia_NameTransfer}
\alias{qingJia_NameTransfer}
\title{处理请假的名称转换}
\usage{
qingJia_NameTransfer(data)
}
\arguments{
\item{data}{deal_qingJia处理后的数据}
}
\value{
返回值
}
\description{
处理请假的名称转换
}
\examples{
qingJia_NameTransfer();
}
