% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/05qingjia_Func.R
\name{qingJia_sum}
\alias{qingJia_sum}
\title{处理请假的汇总数据}
\usage{
qingJia_sum(data, startDate = "2019-07-01", endDate = "2019-07-31")
}
\arguments{
\item{data}{请假明细数据}

\item{startDate}{开始日期}

\item{endDate}{结束日期}
}
\value{
返回值
}
\description{
处理请假的汇总数据
}
\examples{
qingJia_sum();
}
