% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/05qingjia_Func.R
\name{qingJia_sum_byDay}
\alias{qingJia_sum_byDay}
\title{Title}
\usage{
qingJia_sum_byDay(data, startDate = "2019-07-01",
  endDate = "2019-07-31")
}
\arguments{
\item{data}{数据}

\item{startDate}{开始日期}

\item{endDate}{结束日期}
}
\value{
返回值
}
\description{
Title
}
\examples{
qingJia_sum_byDay();
}
