% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03Bwaichu_Func.R
\name{waiChu_NameTransfer}
\alias{waiChu_NameTransfer}
\title{针对外出数据进行名称转换}
\usage{
waiChu_NameTransfer(data)
}
\arguments{
\item{data}{外出数据deal_waiChu处理后结果}
}
\value{
返回值
}
\description{
针对外出数据进行名称转换
}
\examples{
waiChu_NameTransfer();
}
