% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03Bwaichu_Func.R
\name{waiChu_splitIntoMultiRows}
\alias{waiChu_splitIntoMultiRows}
\title{外出数据处理}
\usage{
waiChu_splitIntoMultiRows(data, startDateName = "FStartDate",
  endDateName = "FEndDate")
}
\arguments{
\item{data}{数据}

\item{startDateName}{开始日期}

\item{endDateName}{结束日志}
}
\value{
返回值
}
\description{
外出数据处理
}
\examples{
waiChu_splitIntoMultiRows();
}
