% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03Bwaichu_Func.R
\name{waiChu_sum}
\alias{waiChu_sum}
\title{处理将外出数据进行汇总}
\usage{
waiChu_sum(data, startDate = "2019-07-01", endDate = "2019-07-31")
}
\arguments{
\item{data}{外出明细数据}

\item{startDate}{开始日期}

\item{endDate}{结束日期}
}
\value{
返回值
}
\description{
处理将外出数据进行汇总
}
\examples{
waiChu_sum();
}
