% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03Bwaichu_Func.R
\name{waiChu_sum_byDay}
\alias{waiChu_sum_byDay}
\title{针对外出数据按天进行汇总}
\usage{
waiChu_sum_byDay(data, startDate = "2019-07-01",
  endDate = "2019-07-31")
}
\arguments{
\item{data}{外出数据}

\item{startDate}{开始日期}

\item{endDate}{结束日期}
}
\value{
返回值
}
\description{
针对外出数据按天进行汇总
}
\examples{
waiChu_sum_byDay();
}
